CDN SDK Auto Released By xiaoyan.yan,Version：1.51.23

+++ services/cdn/struct_top_domain.go
@@ -17,11 +17,11 @@ package cdn
 
 // TopDomain is a nested struct in cdn response
 type TopDomain struct {
-	DomainName     string `json:"DomainName" xml:"DomainName"`
-	Rank           int    `json:"Rank" xml:"Rank"`
-	TotalTraffic   string `json:"TotalTraffic" xml:"TotalTraffic"`
-	TrafficPercent string `json:"TrafficPercent" xml:"TrafficPercent"`
-	MaxBps         int    `json:"MaxBps" xml:"MaxBps"`
-	MaxBpsTime     string `json:"MaxBpsTime" xml:"MaxBpsTime"`
-	TotalAccess    int    `json:"TotalAccess" xml:"TotalAccess"`
+	DomainName     string  `json:"DomainName" xml:"DomainName"`
+	Rank           int     `json:"Rank" xml:"Rank"`
+	TotalTraffic   string  `json:"TotalTraffic" xml:"TotalTraffic"`
+	TrafficPercent string  `json:"TrafficPercent" xml:"TrafficPercent"`
+	MaxBps         float64 `json:"MaxBps" xml:"MaxBps"`
+	MaxBpsTime     string  `json:"MaxBpsTime" xml:"MaxBpsTime"`
+	TotalAccess    int     `json:"TotalAccess" xml:"TotalAccess"`
 }
